refactor: organize files in server directory in the appropriate folders
