+++ server/src/constants/index.ts
@@ -1,0 +1,4 @@
+import { manager } from "./manager";
+import config from "./config";
+
+export { manager, config };
