+++ server/src/constants/manager.ts
@@ -1,0 +1,3 @@
+import { RoomManager } from "@/models";
+
+export const manager = new RoomManager();
