+++ server/src/middlewares/index.ts
@@ -1,0 +1,3 @@
+import socketAuth from "./socketAuth";
+
+export { socketAuth };
