+++ server/src/models/index.ts
@@ -1,0 +1,5 @@
+import { Player } from "./Player";
+import { Room } from "./Room";
+import { RoomManager } from "./RoomManager";
+
+export { Player, Room, RoomManager };
