+++ server/src/routes/index.ts
@@ -1,0 +1,4 @@
+import socketRoutes from "./socketRoutes";
+import webRouter from "./webRoutes";
+
+export { socketRoutes, webRouter };
